+++ interview-aissistant/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "ant-design-vue": "^4.0.0-rc.6",
+    "pinia": "^3.0.2",
     "vue": "^3.5.13"
   },
   "devDependencies": {
